requirements.txt
alembic==1.9.0
Babel==2.11.0
bcrypt==4.0.1
bleach==6.0.0
blinker==1.5
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
flake8==6.0.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mailman==0.3.0
Flask-Migrate==4.0.0
Flask-Principal==0.4.0
Flask-Security-Too==5.1.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
mkdocs-material-extensions==1.1.1
passlib==1.7.4
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.7
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.45
SQLAlchemy-Utils==0.39.0
webencodings==0.5.1
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
